requirements.txt
aiohttp<3.8.0,>=3.6.0
PyNaCl<1.5,>=1.3.0
typing-extensions>=3.6.5
yarl<2.0,>=1.0
chardet<5.0,>=2.0
async-timeout<4.0,>=3.0
multidict<7.0,>=4.5
attrs>=17.3.0
discord.py==2.0.0a0
cffi>=1.4.1
pycparser
idna>=2.0